tests: Move `get_channel_name_arg()` helper to `src/test.rs`
In preparation for adding `_spawn()` test variants for the tests in
`src/test.rs` as well, we need to move the helper there, so it can be
used in both `src/test.rs` and `src/platform/test.rs` -- just like the
`fork()` helpers.
